refactor: separate services, fix from mentors comments
