+++ import-service/src/utils/lambda.ts
@@ -1,0 +1,13 @@
+import { APIGatewayProxyResult } from "aws-lambda";
+
+export const generateHttpResponse = <T>(statusCode: number, response: T): APIGatewayProxyResult => {
+  
+    return {
+        statusCode: statusCode,
+        headers: {
+            "Access-Control-Allow-Origin": "*",
+        },
+        body: JSON.stringify(response),
+    };
+};
+
