+++ product-service/src/data/stockForLoad.ts
@@ -1,0 +1,16 @@
+import { productsDB } from "./productsForLoad";
+
+export const stockDB = [
+    {
+        product_id: productsDB[0].id,
+        count: 4,
+    },
+    {
+        product_id: productsDB[1].id,
+        count: 5,
+    },
+    {
+        product_id: productsDB[2].id,
+        count: 2,
+    }
+];
