+++ product-service/src/types/http.ts
@@ -1,0 +1,7 @@
+export enum StatusCode {
+    OK = 200,
+    BAD_REQUEST = 400,
+    FORBIDDEN = 403,
+    NOT_FOUND = 404,
+    SERVER_ERROR = 500
+};
