+++ product-service/src/types/products.ts
@@ -1,0 +1,12 @@
+export interface IProduct {
+    id: string;
+    title: string;
+    price: number;
+    description: string;
+};
+  
+export interface IStock {
+    product_id: string;
+    count: number;
+};
+
