--- src/types/products.ts
@@ -1,12 +1,0 @@
-export interface IProduct {
-      id: string;
-      title: string;
-      price: number;
-      description: string;
-  };
-  
-export interface IStock {
-      product_id: string;
-      count: number;
-  };
-
